+++ src/main/java/com/flint/flint/member/domain/main/Policy.java
@@ -1,7 +1,7 @@
-package com.flint.flint.member.domain;
+package com.flint.flint.member.domain.main;
 
 import com.flint.flint.common.BaseTimeEntity;
-import com.flint.flint.member.spec.Agree;
+import com.flint.flint.member.domain.spec.Agree;
 import com.flint.flint.security.auth.dto.RegisterRequest;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
