+++ src/main/java/com/flint/flint/member/domain/report/PostCommentReport.java
@@ -2,7 +2,7 @@ package com.flint.flint.member.domain.report;
 
 import com.flint.flint.common.BaseTimeEntity;
 import com.flint.flint.community.domain.PostComment;
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
