+++ src/main/java/com/flint/flint/member/domain/scrap/ScrapClub.java
@@ -2,7 +2,7 @@ package com.flint.flint.member.domain.scrap;
 
 import com.flint.flint.club.domain.main.Club;
 import com.flint.flint.common.BaseTimeEntity;
-import com.flint.flint.member.domain.Member;
+import com.flint.flint.member.domain.main.Member;
 import jakarta.persistence.*;
 import lombok.Builder;
 import lombok.Getter;
